+++ Private/Classes/Step.ps1
@@ -10,6 +10,7 @@ class Step {
     [bool]$ContinueOnError = $false
     [datetime]$StartTime
     [Nullable[datetime]]$EndTime
+    [TimeSpan]$Duration = [TimeSpan]::Zero
 
     Step([string]$Name, [Step]$ParentStep = $null, [bool]$ContinueOnError = $false) {
         $this.Name = $Name
@@ -23,5 +24,10 @@ class Step {
         if ($ParentStep) {
             $ParentStep.Children.Add($this)
         }
+    }
+
+    [string] ToString() {
+        $dur = if ($this.EndTime) { ($this.EndTime - $this.StartTime) } else { [TimeSpan]::Zero }
+        return "{0} [{1}] L{2} ({3:c})" -f $this.Name, $this.Status, $this.Level, $dur
     }
 }